move core code for checking paths to CruiseDesignFileContext
